+++ src/multi_vector_simulator/utils/__init__.py
@@ -41,6 +41,111 @@ from .constants_json_strings import (
 from .exceptions import MissingParameterError
 
 
+class ParameterDocumentation:
+    """Helper to access a parameter's information given its variable name"""
+
+    def __init__(
+        self, param_info_file, label_header="label",
+    ):
+        self.param_doc = pd.read_csv(param_info_file).set_index(label_header)
+        self.label_hdr = label_header
+        self.fname = param_info_file
+        self.param_format = {"numeric": float, "str": str, "boolean": bool}
+
+    @property
+    def where_to_find_param_documentation(self):
+        return (
+            "*" * 5
+            + f" Note: The documentation about each of the MVS parameters can be found in the csv file {self.fname}. "
+            + "*" * 5
+        )
+
+    def __get_doc_parameter_info(self, param_label, column_name):
+        """Search the value of a parameter information in the parameter doc
+
+        Parameters
+        ----------
+        param_label: str
+            name of the variable as referenced in the column "label" of the
+            documentation csv file
+        column_name:
+            name of the documentation csv file's column corresponding to the
+            desired information about the parameter
+
+        Returns
+        -------
+        str: value of the given parameter information
+        """
+        if isinstance(param_label, list):
+            answer = []
+            for p_name in param_label:
+                answer.append(self.__get_doc_parameter_info(p_name, column_name))
+        else:
+            try:
+                answer = self.param_doc.loc[param_label][column_name]
+            except KeyError as e:
+                raise KeyError(
+                    f"Either {param_label} is not part of the {self.label_hdr} column of the file {self.fname}, or the column {column_name} does not exist in this file"
+                ).with_traceback(e.__traceback__)
+        return answer
+
+    def get_doc_verbose(self, param_label):
+        answer = self.__get_doc_parameter_info(param_label, "verbose")
+        answer_is_list = True
+        if not isinstance(param_label, list):
+            answer_is_list = False
+            answer = [answer]
+            param_label = [param_label]
+
+        for i in range(len(answer)):
+            if answer[i] == "None":
+                answer[i] = param_label[i].replace("_", " ").title()
+        if answer_is_list is False:
+            answer = answer[0]
+        return answer
+
+    def get_doc_definition(self, param_label):
+        return self.__get_doc_parameter_info(param_label, ":Definition:")
+
+    def get_doc_default(self, param_label):
+        answer = self.__get_doc_parameter_info(param_label, ":Default:")
+        param_type = self.get_doc_type(param_label)
+        if answer == "None":
+            answer = None
+        else:
+            answer = self.param_format[param_type](answer)
+        return answer
+
+    def get_doc_unit(self, param_label):
+        return self.__get_doc_parameter_info(param_label, ":Unit:")
+
+    def get_doc_type(self, param_label):
+        return self.__get_doc_parameter_info(param_label, ":Type:")
+
+
+try:
+    mvs_parameter_file = "MVS_parameters_list.csv"
+    DOC_PATH = os.path.join(
+        os.path.dirname(
+            os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+        ),
+        "docs",
+    )
+    if os.path.exists(DOC_PATH):
+        PARAMETERS_DOC = ParameterDocumentation(
+            param_info_file=os.path.join(DOC_PATH, mvs_parameter_file)
+        )
+    elif os.path.exists(PACKAGE_DATA_PATH):
+        PARAMETERS_DOC = ParameterDocumentation(
+            param_info_file=os.path.join(PACKAGE_DATA_PATH, mvs_parameter_file)
+        )
+    else:
+        PARAMETERS_DOC = None
+
+except FileNotFoundError:
+    PARAMETERS_DOC = None
+
+
 def find_json_input_folders(
     path, specific_file_name=JSON_FNAME, ignore_folders=(OUTPUT_FOLDER,)
 ):
@@ -247,6 +352,17 @@ def compare_input_parameters_with_reference(
                                     + " which can influence the results."
                                     + "In the next release, this parameter will required."
                                 )
+                        elif set_default is True:
+
+                            main_parameters[mp][k][sp] = {
+                                VALUE: PARAMETERS_DOC.get_doc_default(sp),
+                                UNIT: PARAMETERS_DOC.get_doc_unit(sp),
+                            }
+                            logging.warning(
+                                f"You are not providing a value for the parameter '{sp}' of asset '{k}' in asset group '{mp}'"
+                                + f"This parameter is then set to it's default value ({PARAMETERS_DOC.get_doc_default(sp)}).\n"
+                                + PARAMETERS_DOC.where_to_find_param_documentation
+                            )
                         else:
                             # the sub parameter is not provided but is required --> missing
                             param_list = missing_parameters.get(mp, [])